+++ CHANGELOG.md
@@ -7,7 +7,12 @@
 - Support [EIP-1191](https://eips.ethereum.org/EIPS/eip-1191): Add chain id to mixed-case checksum address encoding
 - Add [EIP-4844](https://eips.ethereum.org/EIPS/eip-4844) transaction support
 - Add [EIP-2930](https://eips.ethereum.org/EIPS/eip-2930) transaction support
-- Add `Ethers.blob_fee_per_gas/1` to calculate blob fee per gas
+- Add `Ethers.blob_base_fee/1` to calculate blob fee per gas
+
+### Bug Fixes
+
+- Fix encoding of transactions without any input data
+- Encode `access_list` and `blob_versioned_hashed` correctly in transactions
 
 ## v0.6.1 (2025-01-02)